Removed ScreenCoord and WorldCoord because in bevy Transform is a world coord
